issue-5: paritially fix RSQLite update bugs
In the development version (as of 160722) of RSQLite  trying to
close a connection after calling dbSendQuery causes a warning asking
the user to close the result set. Adding a dbClearResults in
tcplAppend to fix the issue.

Also, the new RSQLite package does not export dbGetQuery, so the
call to dbGetQuery in tcplQuery is changed to DBI::dbGetQuery as
suggested by the RSQLite developers. Likewise, the NAMESPACE is
updated to no longer import dbGetQuery from RSQLite.

Have not yet addressed the class issue when loading a table with
NULL values.